% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cgm_plot}
\alias{cgm_plot}
\title{Plot glucose values over time}
\usage{
cgm_plot(x, times, unit = "mg")
}
\arguments{
\item{x}{vector of glucose readings}

\item{times}{vector of corresponding times, in minutes}

\item{unit}{"mg" if the units are mg/dL or "mmol" if the units are mmol/L. Null value is "mg".}
}
\value{
A plot of glucose values over time.
}
\description{
Plot glucose values over time
}
\examples{
cgm_plot(x=c(rep(100, 10), rep(120, 10), 105, 85), times=seq(0, 1260, 60), unit='mg')
}
