% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{conga}
\alias{conga}
\title{Calculate continuous overall net glycemic action (CONGA)}
\usage{
conga(x, times, n = 1, s = 1, method = "manuscript")
}
\arguments{
\item{x}{vector of glucose readings}

\item{times}{vector of corresponding times, in minutes}

\item{n}{number of hours between "partner" observations. Null value is 1.}

\item{s}{number of minutes of slack used when searching for partners. Null value is 1.}

\item{method}{"manuscript" or "easy". Null value is "manuscript".}
}
\value{
The numeric CONGA value for a given dataset of glucose measurements and times.
}
\description{
Calculate continuous overall net glycemic action (CONGA)
}
\examples{
conga(x=c(rep(100, 10), rep(120, 10), 105, 85), times=seq(0, 1260, 60),
n=1, s=1, method="manuscript")
}
