% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{diff_plot}
\alias{diff_plot}
\title{Plots glucose changes over time}
\usage{
diff_plot(x, times, n = 1, s = 1, unit = "mg")
}
\arguments{
\item{x}{vector of glucose readings}

\item{times}{vector of corresponding times, in minutes}

\item{n}{number of hours between "partner" observations. Null value is 1.}

\item{s}{number of minutes of slack used when searching for partners. Null value is 1.}

\item{unit}{"mg" if the units are mg/dL or "mmol" if the units are mmol/L. Null value is "mg".}
}
\value{
A plot of n-hour glucose differences over time.
}
\description{
Plots glucose changes over time
}
\examples{
diff_plot(x=c(rep(100, 10), rep(120, 10), 105, 85), times=seq(0, 1260, 60), n=1, s=1, unit='mg')
}
