% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{dist_travelled}
\alias{dist_travelled}
\title{Calculate distance travelled}
\usage{
dist_travelled(x)
}
\arguments{
\item{x}{vector of glucose readings}
}
\value{
The numeric distance travelled value for a given dataset of glucose measurements.
}
\description{
Calculate distance travelled
}
\examples{
dist_travelled(x=c(rep(100, 10), rep(120, 10), 105, 85))
}
