% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{gmi}
\alias{gmi}
\title{Calculate Glucose Management Indicator (GMI)}
\usage{
gmi(x, unit = "mg")
}
\arguments{
\item{x}{vector of glucose readings}

\item{unit}{"mg" if the units are mg/dL or "mmol" if the units are mmol/L. Null value is "mg".}
}
\value{
The numeric GMI value for a given dataset of glucose measurements.
}
\description{
Calculate Glucose Management Indicator (GMI)
}
\examples{
gmi(x=c(rep(100, 10), rep(120, 10), 105, 85), unit='mg')
}
