% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{li}
\alias{li}
\title{Calculate the lability index (LI)}
\usage{
li(x, times, k = 60, s = 1)
}
\arguments{
\item{x}{vector of glucose readings}

\item{times}{vector of corresponding times, in minutes}

\item{k}{length of time (in minutes) used to find partners. Null value is 60.}

\item{s}{number of minutes of slack used when searching for partners. Null value is 1.}
}
\value{
The numeric value of the lability index for a given dataset of glucose measurements and times.
}
\description{
Calculate the lability index (LI)
}
\examples{
li(x=c(rep(100, 10), rep(120, 10), 105, 85), times=seq(0, 1260, 60), k=60, s=1)
}
