% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mage}
\alias{mage}
\title{Calculate Mean Amplitude of Glycemic Excursions (MAGE)}
\usage{
mage(x, times)
}
\arguments{
\item{x}{vector of glucose readings}

\item{times}{vector of corresponding times, in minutes}
}
\value{
The numeric MAGE value for a given dataset of glucose measurements and times.
}
\description{
Calculate Mean Amplitude of Glycemic Excursions (MAGE)
}
\examples{
mage(x=c(rep(100, 10), rep(120, 10), 105, 85), times=seq(0, 1260, 60))
}
