% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{modd}
\alias{modd}
\title{Calculate Mean of Daily Differences (MODD)}
\usage{
modd(x, times, s = 1, method = "manuscript")
}
\arguments{
\item{x}{vector of glucose readings}

\item{times}{vector of corresponding times, in minutes}

\item{s}{number of minutes of slack used when searching for partners. Null value is 1.}

\item{method}{"manuscript" or "easy". Null value is "manuscript".}
}
\value{
The numeric MODD value for a given dataset of glucose measurements and times.
}
\description{
Calculate Mean of Daily Differences (MODD)
}
\examples{
modd(x=c(rep(100, 10), rep(120, 10), 105, 85), times=seq(0, 1260, 60), s=1, method='manuscript')
}
