% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{time_on}
\alias{time_on}
\title{Calculate amount of time that the CGM was active}
\usage{
time_on(times, gap = 5)
}
\arguments{
\item{times}{vector of corresponding times, in minutes}

\item{gap}{typical gap between CGM measurements, in minutes. Default is 5}
}
\value{
The numeric amount of time that the CGM device was active in a given dataset.
}
\description{
Calculate amount of time that the CGM was active
}
\examples{
time_on(times=seq(0, 1260, 60), gap=5)
}
