% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{tir}
\alias{tir}
\title{Calculate time in range (TIR)}
\usage{
tir(x, low = 70, high = 180)
}
\arguments{
\item{x}{vector of glucose readings}

\item{low}{lower bound of the range. Default is 70}

\item{high}{upper bound of the range. Default is 180}
}
\value{
The numeric time in range value for a given dataset of glucose measurements and times.
}
\description{
Calculate time in range (TIR)
}
\examples{
tir(x=c(rep(100, 10), rep(120, 10), 105, 85), low=70, high=80)
}
